+++ packages/core/demo/data/index.ts
@@ -27,7 +27,7 @@ import {
 	createAngularChartApp,
 	createVueChartApp,
 	createVanillaChartApp,
-	createSvelteChartApp,
+	createSvelteChartApp
 } from "../create-codesandbox";
 
 import { Tools } from "@carbon/charts/tools";
@@ -36,153 +36,150 @@ export const chartTypes = {
 	AreaChart: {
 		vanilla: "AreaChart",
 		angular: "ibm-simple-bar-chart",
-		vue: "ccv-simple-bar-chart",
+		vue: "ccv-simple-bar-chart"
 	},
 	SimpleBarChart: {
 		vanilla: "SimpleBarChart",
 		angular: "ibm-simple-bar-chart",
-		vue: "ccv-simple-bar-chart",
+		vue: "ccv-simple-bar-chart"
 	},
 	GroupedBarChart: {
 		vanilla: "GroupedBarChart",
 		angular: "ibm-grouped-bar-chart",
-		vue: "ccv-grouped-bar-chart",
+		vue: "ccv-grouped-bar-chart"
 	},
 	StackedBarChart: {
 		vanilla: "StackedBarChart",
 		angular: "ibm-stacked-bar-chart",
-		vue: "ccv-stacked-bar-chart",
+		vue: "ccv-stacked-bar-chart"
 	},
 	BubbleChart: {
 		vanilla: "BubbleChart",
 		angular: "ibm-bubble-chart",
-		vue: "ccv-bubble-chart",
+		vue: "ccv-bubble-chart"
 	},
 	LineChart: {
 		vanilla: "LineChart",
 		angular: "ibm-line-chart",
-		vue: "ccv-line-chart",
+		vue: "ccv-line-chart"
 	},
 	ScatterChart: {
 		vanilla: "ScatterChart",
 		angular: "ibm-scatter-chart",
-		vue: "ccv-scatter-chart",
+		vue: "ccv-scatter-chart"
 	},
 	PieChart: {
 		vanilla: "PieChart",
 		angular: "ibm-pie-chart",
-		vue: "ccv-pie-chart",
+		vue: "ccv-pie-chart"
 	},
 	GaugeChart: {
 		vanilla: "GaugeChart",
 		angular: "ibm-gauge-chart",
-		vue: "ccv-gauge-chart",
+		vue: "ccv-gauge-chart"
 	},
 	DonutChart: {
 		vanilla: "DonutChart",
 		angular: "ibm-donut-chart",
-		vue: "ccv-donut-chart",
+		vue: "ccv-donut-chart"
 	},
 	RadarChart: {
 		vanilla: "RadarChart",
 		angular: "ibm-radar-chart",
-		vue: "ccv-radar-chart",
-	},
+		vue: "ccv-radar-chart"
+	}
 };
 
 let allDemoGroups = [
 	{
 		title: "Area",
+		description: "Area charts are similar to line charts, but the areas below the lines are filled with colors or patterns. Stacked charts are useful for comparing proportional contributions within a category. They plot the relative value that each data series contributes to the total.",
 		demos: [
 			{
 				options: areaDemos.areaTimeSeriesCurvedOptions,
 				data: areaDemos.areaTimeSeriesCurvedData,
-				chartType: chartTypes.AreaChart,
-				isDemoExample: true,
+				chartType: chartTypes.AreaChart
 			},
 			{
 				options: areaDemos.areaTimeSeriesOptions,
 				data: areaDemos.areaTimeSeriesData,
-				chartType: chartTypes.AreaChart,
-				isDemoExample: true,
-			},
-		],
+				chartType: chartTypes.AreaChart
+			}
+		]
 	},
 	{
 		title: "Bar (vertical)",
+		description: "Bar charts use vertical or horizontal data markers to compare individual values. You can use them to compare discrete data or show trends over time.",
 		demos: [
 			{
 				options: barDemos.simpleBarOptions,
 				data: barDemos.simpleBarData,
-				chartType: chartTypes.SimpleBarChart,
-				isDemoExample: true,
+				chartType: chartTypes.SimpleBarChart
 			},
 			{
 				options: barDemos.simpleBarTimeSeriesOptions,
 				data: barDemos.simpleBarTimeSeriesData,
-				chartType: chartTypes.SimpleBarChart,
-				isDemoExample: true,
+				chartType: chartTypes.SimpleBarChart
 			},
 			{
 				options: barDemos.simpleBarFixedDomainOptions,
 				data: barDemos.simpleBarData,
-				chartType: chartTypes.SimpleBarChart,
+				chartType: chartTypes.SimpleBarChart
 			},
 			{
 				options: barDemos.simpleBarEmptyStateOptions,
 				data: barDemos.simpleBarEmptyStateData,
 				chartType: chartTypes.SimpleBarChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: barDemos.simpleBarSkeletonOptions,
 				data: barDemos.simpleBarSkeletonData,
 				chartType: chartTypes.SimpleBarChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
+				description: "A grouped bar chart, also known as a clustered bar graph, multi-set bar chart, or grouped column chart, is a type of bar graph that is used to compare values across multiple categories.",
 				options: barDemos.groupedBarOptions,
 				data: barDemos.groupedBarData,
-				chartType: chartTypes.GroupedBarChart,
-				isDemoExample: true,
+				chartType: chartTypes.GroupedBarChart
 			},
 			{
 				options: barDemos.groupedBarEmptyStateOptions,
 				data: barDemos.groupedBarEmptyStateData,
 				chartType: chartTypes.GroupedBarChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: barDemos.groupedBarSkeletonOptions,
 				data: barDemos.groupedBarSkeletonData,
 				chartType: chartTypes.GroupedBarChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
+				description: "Stacked bar charts are useful for comparing proportional contributions within a category. They plot the relative value that each data series contributes to the total.",
 				options: barDemos.stackedBarOptions,
 				data: barDemos.stackedBarData,
-				chartType: chartTypes.StackedBarChart,
-				isDemoExample: true,
+				chartType: chartTypes.StackedBarChart
 			},
 			{
 				options: barDemos.stackedBarTimeSeriesOptions,
 				data: barDemos.stackedBarTimeSeriesData,
-				chartType: chartTypes.StackedBarChart,
-				isDemoExample: true,
+				chartType: chartTypes.StackedBarChart
 			},
 			{
 				options: barDemos.stackedBarEmptyStateOptions,
 				data: barDemos.stackedBarEmptyStateData,
 				chartType: chartTypes.StackedBarChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: barDemos.stackedBarSkeletonOptions,
 				data: barDemos.stackedBarSkeletonData,
 				chartType: chartTypes.StackedBarChart,
-				isDemoExample: true,
-			},
-		],
+				isDemoExample: false
+			}
+		]
 	},
 	{
 		title: "Bar (horizontal)",
@@ -190,105 +187,98 @@ let allDemoGroups = [
 			{
 				options: barDemos.simpleHorizontalBarOptions,
 				data: barDemos.simpleHorizontalBarData,
-				chartType: chartTypes.SimpleBarChart,
-				isDemoExample: true,
+				chartType: chartTypes.SimpleBarChart
 			},
 			{
 				options: barDemos.simpleHorizontalBarTimeSeriesOptions,
 				data: barDemos.simpleHorizontalBarTimeSeriesData,
-				chartType: chartTypes.SimpleBarChart,
-				isDemoExample: true,
+				chartType: chartTypes.SimpleBarChart
 			},
 			{
 				options: barDemos.simpleHorizontalBarEmptyStateOptions,
 				data: barDemos.simpleHorizontalBarEmptyStateData,
 				chartType: chartTypes.SimpleBarChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: barDemos.simpleHorizontalBarSkeletonOptions,
 				data: barDemos.simpleHorizontalBarSkeletonData,
 				chartType: chartTypes.SimpleBarChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: barDemos.groupedHorizontalBarOptions,
 				data: barDemos.groupedHorizontalBarData,
-				chartType: chartTypes.GroupedBarChart,
-				isDemoExample: true,
+				chartType: chartTypes.GroupedBarChart
 			},
 			{
 				options: barDemos.groupedHorizontalBarEmptyStateOptions,
 				data: barDemos.groupedHorizontalBarEmptyStateData,
 				chartType: chartTypes.GroupedBarChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: barDemos.groupedHorizontalBarSkeletonOptions,
 				data: barDemos.groupedHorizontalBarSkeletonData,
 				chartType: chartTypes.GroupedBarChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: barDemos.stackedHorizontalBarOptions,
 				data: barDemos.stackedHorizontalBarData,
-				chartType: chartTypes.StackedBarChart,
-				isDemoExample: true,
+				chartType: chartTypes.StackedBarChart
 			},
 			{
 				options: barDemos.stackedHorizontalBarTimeSeriesOptions,
 				data: barDemos.stackedHorizontalBarTimeSeriesData,
-				chartType: chartTypes.StackedBarChart,
-				isDemoExample: true,
+				chartType: chartTypes.StackedBarChart
 			},
 			{
 				options: barDemos.stackedHorizontalBarEmptyStateOptions,
 				data: barDemos.stackedHorizontalBarEmptyStateData,
 				chartType: chartTypes.StackedBarChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: barDemos.stackedHorizontalBarSkeletonOptions,
 				data: barDemos.stackedHorizontalBarSkeletonData,
 				chartType: chartTypes.StackedBarChart,
-				isDemoExample: true,
-			},
-		],
+				isDemoExample: false
+			}
+		]
 	},
 	{
 		title: "Bubble",
+		description: "Bubble charts use data points and bubbles to plot measures anywhere along a scale. One measure is plotted along each axis. The size of the bubble represents the third measure. You can use bubble charts to represent financial data or any data where measured values are related.",
 		demos: [
 			{
 				options: bubbleDemos.bubbleDoubleLinearOptions,
 				data: bubbleDemos.bubbleDoubleLinearData,
-				chartType: chartTypes.BubbleChart,
-				isDemoExample: true,
+				chartType: chartTypes.BubbleChart
 			},
 			{
 				options: bubbleDemos.bubbleTimeSeriesOptions,
 				data: bubbleDemos.bubbleTimeSeriesData,
-				chartType: chartTypes.BubbleChart,
-				isDemoExample: true,
+				chartType: chartTypes.BubbleChart
 			},
 			{
 				options: bubbleDemos.bubbleDiscreteOptions,
 				data: bubbleDemos.bubbleDiscreteData,
-				chartType: chartTypes.BubbleChart,
-				isDemoExample: true,
+				chartType: chartTypes.BubbleChart
 			},
 			{
 				options: bubbleDemos.bubbleEmptyStateOptions,
 				data: bubbleDemos.bubbleEmptyStateData,
 				chartType: chartTypes.BubbleChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: bubbleDemos.bubbleSkeletonOptions,
 				data: bubbleDemos.bubbleSkeletonData,
 				chartType: chartTypes.BubbleChart,
-				isDemoExample: true,
-			},
-		],
+				isDemoExample: false
+			}
+		]
 	},
 	{
 		title: "Donut",
@@ -296,74 +286,77 @@ let allDemoGroups = [
 			{
 				options: donutDemos.donutOptions,
 				data: donutDemos.donutData,
-				chartType: chartTypes.DonutChart,
-				isDemoExample: true,
+				chartType: chartTypes.DonutChart
 			},
 			{
 				options: donutDemos.donutEmptyStateOptions,
 				data: donutDemos.donutEmptyStateData,
 				chartType: chartTypes.DonutChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: donutDemos.donutSkeletonOptions,
 				data: donutDemos.donutSkeletonData,
 				chartType: chartTypes.DonutChart,
-				isDemoExample: true,
-			},
-		],
+				isDemoExample: false
+			}
+		]
 	},
 	{
 		title: "Line",
+		description: "Line charts plot data at regular intervals connected by lines. You can use line visualizations to show trends over time and compare several data sets.",
 		demos: [
 			{
 				options: lineDemos.lineTimeSeriesOptions,
 				data: lineDemos.lineTimeSeriesData,
-				chartType: chartTypes.LineChart,
-				isDemoExample: true,
+				chartType: chartTypes.LineChart
 			},
 			{
 				options: lineDemos.lineTimeSeriesDenseOptions,
 				data: lineDemos.lineTimeSeriesDenseData,
-				chartType: chartTypes.LineChart,
+				chartType: chartTypes.LineChart
 			},
 			{
 				options: lineDemos.lineOptions,
 				data: lineDemos.lineData,
-				chartType: chartTypes.LineChart,
-				isDemoExample: true,
+				chartType: chartTypes.LineChart
 			},
 			{
 				options: lineDemos.lineTimeSeriesRotatedTicksOptions,
 				data: lineDemos.lineTimeSeriesDataRotatedTicks,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				options: lineDemos.lineTimeSeriesHorizontalOptions,
 				data: lineDemos.lineTimeSeriesData,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				options: lineDemos.lineHorizontalOptions,
 				data: lineDemos.lineData,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				options: lineDemos.lineTimeSeriesWithThresholdsOptions,
 				data: lineDemos.lineTimeSeriesData,
-				chartType: chartTypes.LineChart,
+				chartType: chartTypes.LineChart
 			},
 			{
 				options: lineDemos.lineEmptyStateOptions,
 				data: lineDemos.lineEmptyStateData,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				options: lineDemos.lineSkeletonOptions,
 				data: lineDemos.lineSkeletonData,
 				chartType: chartTypes.LineChart,
-			},
-		],
+				isDemoExample: false
+			}
+		]
 	},
 	{
 		title: "Pie",
@@ -371,22 +364,21 @@ let allDemoGroups = [
 			{
 				options: pieDemos.pieOptions,
 				data: pieDemos.pieData,
-				chartType: chartTypes.PieChart,
-				isDemoExample: true,
+				chartType: chartTypes.PieChart
 			},
 			{
 				options: pieDemos.pieEmptyStateOptions,
 				data: pieDemos.pieEmptyStateData,
 				chartType: chartTypes.PieChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: pieDemos.pieSkeletonOptions,
 				data: pieDemos.pieSkeletonData,
 				chartType: chartTypes.PieChart,
-				isDemoExample: true,
-			},
-		],
+				isDemoExample: false
+			}
+		]
 	},
 	{
 		title: "Gauge",
@@ -395,79 +387,76 @@ let allDemoGroups = [
 				options: gaugeDemos.gaugeOptionsSemi,
 				data: gaugeDemos.gaugeData,
 				chartType: chartTypes.GaugeChart,
-				isDemoExample: true,
+				isDemoExample: true
 			},
 			{
 				options: gaugeDemos.gaugeOptionsCircular,
 				data: gaugeDemos.gaugeData,
 				chartType: chartTypes.GaugeChart,
-				isDemoExample: true,
-			},
-		],
+				isDemoExample: true
+			}
+		]
 	},
 	{
 		title: "Scatter",
+		description: "Scatter plot visualizations use data points to plot two measures anywhere along a scale, not only at regular tick marks. You can use scatter plots to explore correlations between different measures.",
 		demos: [
 			{
 				options: scatterDemos.doubleLinearScatterOptions,
 				data: scatterDemos.doubleLinearScatterData,
-				chartType: chartTypes.ScatterChart,
-				isDemoExample: true,
+				chartType: chartTypes.ScatterChart
 			},
 			{
 				options: scatterDemos.scatterTimeSeriesOptions,
 				data: scatterDemos.scatterTimeSeriesData,
-				chartType: chartTypes.ScatterChart,
-				isDemoExample: true,
+				chartType: chartTypes.ScatterChart
 			},
 			{
 				options: scatterDemos.scatterDiscreteOptions,
 				data: scatterDemos.scatterDiscreteData,
-				chartType: chartTypes.ScatterChart,
-				isDemoExample: true,
+				chartType: chartTypes.ScatterChart
 			},
 			{
 				options: scatterDemos.scatterEmptyStateOptions,
 				data: scatterDemos.scatterEmptyStateData,
 				chartType: chartTypes.ScatterChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: scatterDemos.scatterSkeletonOptions,
 				data: scatterDemos.scatterSkeletonData,
 				chartType: chartTypes.ScatterChart,
-				isDemoExample: true,
-			},
-		],
+				isDemoExample: false
+			}
+		]
 	},
 	{
 		title: "Step",
+		description: "Stepped line charts plot data at regular intervals, forming a series of steps between data points. You can use line visualizations to show trends over time and compare several data sets.",
 		demos: [
 			{
 				options: stepDemos.stepOptions,
 				data: stepDemos.stepData,
-				chartType: chartTypes.LineChart,
-				isDemoExample: true,
+				chartType: chartTypes.LineChart
 			},
 			{
 				options: stepDemos.stepTimeSeriesOptions,
 				data: stepDemos.stepTimeSeriesData,
-				chartType: chartTypes.LineChart,
-				isDemoExample: true,
+				chartType: chartTypes.LineChart
 			},
 			{
 				options: stepDemos.stepEmptyStateOptions,
 				data: stepDemos.stepEmptyStateData,
 				chartType: chartTypes.LineChart,
-				isDemoExample: true,
+				isDemoExample: false
 			},
 			{
 				options: stepDemos.stepSkeletonOptions,
 				data: stepDemos.stepSkeletonData,
 				chartType: chartTypes.LineChart,
-				isDemoExample: true,
-			},
-		],
+				isDemoExample: false
+			}
+		]
 	},
 	{
 		title: "Time series axis",
@@ -476,16 +465,19 @@ let allDemoGroups = [
 				options: timeSeriesAxisDemos.lineTimeSeries15secondsOptions,
 				data: timeSeriesAxisDemos.lineTimeSeriesData15seconds,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				options: timeSeriesAxisDemos.lineTimeSeriesMinuteOptions,
 				data: timeSeriesAxisDemos.lineTimeSeriesDataMinute,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				options: timeSeriesAxisDemos.lineTimeSeries30minutesOptions,
 				data: timeSeriesAxisDemos.lineTimeSeriesData30minutes,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				options:
@@ -493,6 +485,7 @@ let allDemoGroups = [
 				data:
 					timeSeriesAxisDemos.lineTimeSeriesDataHourlyDefaultTicksFormats,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				options:
@@ -500,16 +493,19 @@ let allDemoGroups = [
 				data:
 					timeSeriesAxisDemos.lineTimeSeriesDataHourlyCustomTicksFormats,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				data: timeSeriesAxisDemos.lineTimeSeriesDataDaily,
 				options: timeSeriesAxisDemos.lineTimeSeriesDailyOptions,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				data: timeSeriesAxisDemos.lineTimeSeriesDataWeekly,
 				options: timeSeriesAxisDemos.lineTimeSeriesWeeklyOptions,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				data:
@@ -517,39 +513,46 @@ let allDemoGroups = [
 				options:
 					timeSeriesAxisDemos.lineTimeSeriesMonthlyDefaultLocaleOptions,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				data: timeSeriesAxisDemos.lineTimeSeriesDataMonthlyCustomLocale,
 				options:
 					timeSeriesAxisDemos.lineTimeSeriesMonthlyCustomLocaleOptions,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				data: timeSeriesAxisDemos.lineTimeSeriesDataQuarterly,
 				options: timeSeriesAxisDemos.lineTimeSeriesQuarterlyOptions,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				data: timeSeriesAxisDemos.lineTimeSeriesDataYearly,
 				options: timeSeriesAxisDemos.lineTimeSeriesYearlyOptions,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				data: timeSeriesAxisDemos.lineTimeSeriesDataSingleDatum,
 				options: timeSeriesAxisDemos.lineTimeSeriesSingleDatumOptions,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				data: timeSeriesAxisDemos.lineTimeSeriesNoExtendedDomainData,
 				options:
 					timeSeriesAxisDemos.lineTimeSeriesNoExtendedDomainOptions,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				data: timeSeriesAxisDemos.lineTimeSeriesDataTwoIdenticalLabels,
 				options:
 					timeSeriesAxisDemos.lineTimeSeriesTwoIdenticalLabelsOptions,
 				chartType: chartTypes.LineChart,
+				isDemoExample: false
 			},
 			{
 				data:
@@ -557,8 +560,9 @@ let allDemoGroups = [
 				options:
 					timeSeriesAxisDemos.lineTimeSeriesAllLabelsInPrimaryFormatOptions,
 				chartType: chartTypes.LineChart,
-			},
-		],
+				isDemoExample: false
+			}
+		]
 	},
 	{
 		title: "Radar",
@@ -566,23 +570,20 @@ let allDemoGroups = [
 			{
 				data: radarDemos.radarData,
 				options: radarDemos.radarOptions,
-				chartType: chartTypes.RadarChart,
-				isDemoExample: true,
+				chartType: chartTypes.RadarChart
 			},
 			{
 				data: radarDemos.radarWithMissingDataData,
 				options: radarDemos.radarWithMissingDataOptions,
-				chartType: chartTypes.RadarChart,
-				isDemoExample: true,
+				chartType: chartTypes.RadarChart
 			},
 			{
 				data: radarDemos.radarDenseData,
 				options: radarDemos.radarDenseOptions,
-				chartType: chartTypes.RadarChart,
-				isDemoExample: true,
-			},
-		],
-	},
+				chartType: chartTypes.RadarChart
+			}
+		]
+	}
 ] as any;
 
 const formatTitleString = (str) =>
@@ -629,8 +630,9 @@ export const storybookDemoGroups = Tools.clone(allDemoGroups);
 
 // in the demo page we want to show only demos with isDemoExample = true
 export const demoGroups = Tools.clone(allDemoGroups)
+	// remove demoGroup if its children don't have any demo examples
+	.filter((demoGroup) => demoGroup.demos.filter(demo => demo.isDemoExample !== false).length > 0)
 	.map((demoGroup) => {
-		demoGroup.demos = demoGroup.demos.filter((demo) => demo.isDemoExample);
+		demoGroup.demos = demoGroup.demos.filter(demo => demo.isDemoExample !== false);
 		return demoGroup;
-	})
-	.filter((demoGroup) => demoGroup.demos.length); // remove demoGroup if it's children are all with isDemoExample = false
+	});